make TOS separate component, add metadata and analytics
