+++ app/board/page.tsx
@@ -3,6 +3,11 @@ import { boardList } from "@/lib/boards";
 import { BoardsView } from "@/components/BoardsView";
 import { Unauthorized } from "@/components/Unauthorized";
 
+export const metadata = {
+  title: "Boards",
+  description: "Boards for everynyan",
+};
+
 export default async function BoardsPage() {
   const loggedIn = await getAuthUser();
   if (!loggedIn) {
