+++ app/create/page.tsx
@@ -2,6 +2,11 @@ import { PostCreator } from "@/components/PostCreator";
 import { Unauthorized } from "@/components/Unauthorized";
 import { getAuthUser } from "@/lib/user";
 
+export const metadata = {
+  title: "Create Post",
+  description: "Create a post on everynyan",
+};
+
 export default async function CreatePost() {
   if (!await getAuthUser()) {
     return <Unauthorized />;
